refactor: move Delay to delay.rs
